pyproject.toml
[project]
name = "psytk-translate"
version = "0.1.0"
description = "A tool to facilitate psytoolkit survey translation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "jinja2>=3.1.6",
    "pyyaml>=6.0.2",
]
